+++ app/jobs/test/thread_time_test_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Test::ThreadTimeTestJob < ApplicationJob
+  def perform
+    puts "Perform 30 seconds job in queue #{queue_name}"
+    sleep(30)
+    puts "ThreadTimeTestJob completed."
+  end
+end